add dev documentation in deduplicator
