+++ deduplicator/src/lib/utils.rs
@@ -1,3 +1,6 @@
+//! Generic utilities.
+
+use std::borrow::Borrow;
 use std::convert::{TryFrom, TryInto};
 use std::ffi::CString;
 use std::path::PathBuf;
@@ -9,29 +12,84 @@ use prog_rs::prelude::*;
 use rusqlite::{Connection, NO_PARAMS};
 use tools::{Address, CompatibleDB};
 
+/// Compare two elements wrapped into an option using provided comparison function. If at least one
+/// of the elements is `None`, this will return false.
+///
+/// # Example
+/// ```
+/// use deduplicator::utils::*;
+///
+/// let same_sign = |x: &i64, y: &i64| (*x < 0) == (*y < 0);
+///
+/// assert!(field_compare(Some(3), Some(42), same_sign));
+/// assert!(field_compare(Some(-1), Some(-10), same_sign));
+///
+/// assert!(!field_compare(Some(-1), Some(1), same_sign));
+/// assert!(!field_compare(None, Some(0), same_sign));
+/// assert!(!field_compare(None, None, same_sign));
+/// ```
 pub fn field_compare<T>(
-    field1: &Option<T>,
-    field2: &Option<T>,
+    field1: impl Borrow<Option<T>>,
+    field2: impl Borrow<Option<T>>,
     compare: impl Fn(&T, &T) -> bool,
 ) -> bool {
-    match (field1.as_ref(), field2.as_ref()) {
+    match (field1.borrow().as_ref(), field2.borrow().as_ref()) {
         (Some(field1), Some(field2)) => compare(field1, field2),
         _ => false,
     }
 }
 
+/// Compare two elements wrapped into an option using provided comparison function. If one of the
+/// elements is `None` and the other is not, this will return false.
+///
+/// # Example
+/// ```
+/// use deduplicator::utils::*;
+///
+/// let same_sign = |x: &i64, y: &i64| (*x < 0) == (*y < 0);
+///
+/// assert!(opt_field_compare(Some(3), Some(42), same_sign));
+/// assert!(opt_field_compare(Some(-1), Some(-10), same_sign));
+/// assert!(opt_field_compare(None, None, same_sign));
+///
+/// assert!(!opt_field_compare(Some(-1), Some(1), same_sign));
+/// assert!(!opt_field_compare(None, Some(0), same_sign));
+/// ```
 pub fn opt_field_compare<T>(
-    field1: &Option<T>,
-    field2: &Option<T>,
+    field1: impl Borrow<Option<T>>,
+    field2: impl Borrow<Option<T>>,
     compare: impl Fn(&T, &T) -> bool,
 ) -> bool {
-    match (field1.as_ref(), field2.as_ref()) {
+    match (field1.borrow().as_ref(), field2.borrow().as_ref()) {
         (None, None) => true,
         (Some(field1), Some(field2)) => compare(field1, field2),
         _ => false,
     }
 }
 
+/// Given an address, return its array reprensation used by libpostal.
+///
+/// # Example
+/// ```
+/// use deduplicator::utils::*;
+/// use std::ffi::CString;
+/// use tools::Address;
+///
+/// let address = Address {
+///     number: Some("54".to_string()),
+///     street: Some("rue des Koubis".to_string()),
+///     city: Some("Paris".to_string()),
+///     ..Address::default()
+/// };
+///
+/// assert_eq!(postal_repr(&Address::default()), vec![]);
+/// assert!(postal_repr(&address).contains(
+///     &rpostal::Address {
+///         label: CString::new("city".as_bytes()).unwrap(),
+///         value: CString::new("Paris".as_bytes()).unwrap(),
+///     }
+/// ));
+/// ```
 pub fn postal_repr(address: &Address) -> Vec<rpostal::Address> {
     [
         ("house_number", &address.number),
@@ -52,6 +110,15 @@ pub fn postal_repr(address: &Address) -> Vec<rpostal::Address> {
     .collect()
 }
 
+/// Check if an SQLite error is a constraint violation.
+///
+/// # Example
+/// ```
+/// use deduplicator::utils::*;
+/// use rusqlite::Connection;
+///
+/// assert!(!is_constraint_violation_error(&Connection::open("file/not/exists.sql").unwrap_err()));
+/// ```
 pub fn is_constraint_violation_error(err: &rusqlite::Error) -> bool {
     match err {
         rusqlite::Error::SqliteFailure(
@@ -65,6 +132,7 @@ pub fn is_constraint_violation_error(err: &rusqlite::Error) -> bool {
     }
 }
 
+/// Load addresses from an SQLite file, into a deduplicator.
 pub fn load_from_sqlite<F, R>(
     deduplication: &mut Deduplicator,
     path: PathBuf,
